graphics update
added a bunch of content on the graphics class. This will need to be
implemented in the main class in the future, but for now I am keeping it
in the graphics class for testing purposes until it is ready for use.
When compiling the graphics, you will two different numbers in the
command line, this is the position of the mouse (X,Y coordinates) and it
is constantly updating.  You should also see a sliding box.